+++ client-modern/src/main/java/com/paraguacraft/pvp/modern/ParaguacraftPvPModern.java
@@ -21,7 +21,7 @@ import net.minecraft.client.MinecraftClient;
 public class ParaguacraftPvPModern implements ClientModInitializer {
 
     public static final String MOD_ID = "paraguacraftpvp-modern";
-    public static final String VERSION = "0.8.1";
+    public static final String VERSION = "0.8.2";
 
     @Override
     public void onInitializeClient() {